Integrtae common and lib files in filters
